Initialisation du projet : arbo dossiers, components et modules
