+++ imports/ui/modules/Account/Inscription/Fields/array.js
@@ -1,0 +1,7 @@
+const fields = [
+  { name: "username",   type: "text",      placeholder: "Username" },
+  { name: "password",   type: "password",  placeholder: "Password"  },
+  { name: "email",      type: "text",      placeholder: "Email"     },
+];
+
+export default fields;
