+++ imports/ui/modules/Account/Verify.jsx
@@ -1,3 +1,3 @@
 import React from 'react';
 
-export default () => <div />
+export default () => <div/>
