--- imports/ui/modules/Inscription/Fields/array.js
@@ -1,7 +1,0 @@
-const fields = [
-  { name: "email",      type: "text",      placeholder: "Email"     },
-  { name: "password",   type: "password",  placeholder: "Password"  },
-  { name: "username",   type: "text",      placeholder: "Username" },
-];
-
-export default fields;
